Local_LM/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
httpx==0.25.2
aiofiles==23.2.1
PyPDF2==3.0.1
python-docx==1.1.0
textract==1.6.5

